ai_ops_assistant/requirements.txt
openai>=2.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.5.0,<3.0.0

